@@ -1,17 +1,66 @@
-/// The Amplify configuration generated by the Amplify CLI.
-///
-/// Replace the placeholder values below with the real configuration output
-/// from `amplify pull` or `amplify init`. Keeping the structure in a dedicated
-/// file allows us to avoid checking in platform-specific JSON assets.
-const String amplifyconfig = '''
-{
-  "auth": {
-    "plugins": {
-      "awsCognitoAuthPlugin": {
-        "UserAgent": "aws-amplify-cli/2.0",
-        "Version": "1.0"
-      }
+const amplifyconfig = '''{
+    "UserAgent": "aws-amplify-cli/2.0",
+    "Version": "1.0",
+    "auth": {
+        "plugins": {
+            "awsCognitoAuthPlugin": {
+                "UserAgent": "aws-amplify-cli/0.1.0",
+                "Version": "0.1.0",
+                "IdentityManager": {
+                    "Default": {}
+                },
+                "CredentialsProvider": {
+                    "CognitoIdentity": {
+                        "Default": {
+                            "PoolId": "us-west-2:6930cd8e-493f-4574-9404-4c60fa29ce6b",
+                            "Region": "us-west-2"
+                        }
+                    }
+                },
+                "CognitoUserPool": {
+                    "Default": {
+                        "PoolId": "us-west-2_3HsrxtPWY",
+                        "AppClientId": "6cklkdk33knsfhj0te2755ji3o",
+                        "Region": "us-west-2"
+                    }
+                },
+                "Auth": {
+                    "Default": {
+                        "OAuth": {
+                            "WebDomain": "coalition-auth-app-dev-coalition.auth.us-west-2.amazoncognito.com",
+                            "AppClientId": "6cklkdk33knsfhj0te2755ji3o",
+                            "SignInRedirectURI": "myapp://auth/",
+                            "SignOutRedirectURI": "myapp://auth/",
+                            "Scopes": [
+                                "phone",
+                                "email",
+                                "openid",
+                                "profile"
+                            ]
+                        },
+                        "authenticationFlowType": "USER_SRP_AUTH",
+                        "socialProviders": [
+                            "GOOGLE"
+                        ],
+                        "usernameAttributes": [],
+                        "signupAttributes": [
+                            "EMAIL"
+                        ],
+                        "passwordProtectionSettings": {
+                            "passwordPolicyMinLength": 8,
+                            "passwordPolicyCharacters": []
+                        },
+                        "mfaConfiguration": "OPTIONAL",
+                        "mfaTypes": [
+                            "SMS",
+                            "TOTP"
+                        ],
+                        "verificationMechanisms": [
+                            "EMAIL"
+                        ]
+                    }
+                }
+            }
+        }
     }
-  }
-}
-''';
+}''';
