@@ -1,8 +1,37 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_riverpod/legacy.dart';
 
-import '../models/user_summary.dart';
 import '../../../providers/app_providers.dart';
 import '../../../services/auth_service.dart';
+import '../models/user_summary.dart';
+
+class AwaitingConfirmationState {
+  const AwaitingConfirmationState({
+    required this.username,
+    required this.email,
+    required this.password,
+    this.deliveryDestination,
+    this.codeSentAt,
+  });
+
+  final String username;
+  final String email;
+  final String password;
+  final String? deliveryDestination;
+  final DateTime? codeSentAt;
+
+  AwaitingConfirmationState copyWith({
+    String? deliveryDestination,
+    DateTime? codeSentAt,
+  }) {
+    return AwaitingConfirmationState(
+      username: username,
+      email: email,
+      password: password,
+      deliveryDestination: deliveryDestination ?? this.deliveryDestination,
+      codeSentAt: codeSentAt ?? this.codeSentAt,
+    );
+  }
+}
 
 class AuthState {
   const AuthState({
@@ -11,6 +40,8 @@ class AuthState {
     required this.isSignedIn,
     this.user,
     this.errorMessage,
+    this.errorCode,
+    this.awaitingConfirmation,
   });
 
   const AuthState.loading()
@@ -18,13 +49,17 @@ class AuthState {
         initialized = false,
         isSignedIn = false,
         user = null,
-        errorMessage = null;
+        errorMessage = null,
+        errorCode = null,
+        awaitingConfirmation = null;
 
   final bool isLoading;
   final bool initialized;
   final bool isSignedIn;
   final UserSummary? user;
   final String? errorMessage;
+  final String? errorCode;
+  final AwaitingConfirmationState? awaitingConfirmation;
 
   AuthState copyWith({
     bool? isLoading,
@@ -32,13 +67,23 @@ class AuthState {
     bool? isSignedIn,
     UserSummary? user,
     String? errorMessage,
+    bool resetErrorMessage = false,
+    String? errorCode,
+    bool resetErrorCode = false,
+    AwaitingConfirmationState? awaitingConfirmation,
+    bool resetAwaitingConfirmation = false,
   }) {
     return AuthState(
       isLoading: isLoading ?? this.isLoading,
       initialized: initialized ?? this.initialized,
       isSignedIn: isSignedIn ?? this.isSignedIn,
       user: user ?? this.user,
-      errorMessage: errorMessage ?? this.errorMessage,
+      errorMessage:
+          resetErrorMessage ? null : (errorMessage ?? this.errorMessage),
+      errorCode: resetErrorCode ? null : (errorCode ?? this.errorCode),
+      awaitingConfirmation: resetAwaitingConfirmation
+          ? null
+          : (awaitingConfirmation ?? this.awaitingConfirmation),
     );
   }
 }
@@ -59,6 +104,7 @@ class AuthController extends StateNotifier<AuthState> {
     _bootstrapped = true;
     state = state.copyWith(isLoading: true, initialized: false);
     try {
+      await _authService.configureIfNeeded();
       final signedIn = await _authService.isSignedIn();
       UserSummary? user;
       if (signedIn) {
@@ -93,7 +139,12 @@ class AuthController extends StateNotifier<AuthState> {
     required String usernameOrEmail,
     required String password,
   }) async {
-    state = state.copyWith(isLoading: true, errorMessage: null);
+    state = state.copyWith(
+      isLoading: true,
+      resetErrorMessage: true,
+      resetErrorCode: true,
+      resetAwaitingConfirmation: true,
+    );
     try {
       await _authService.signInEmail(
         usernameOrEmail: usernameOrEmail,
@@ -107,11 +158,15 @@ class AuthController extends StateNotifier<AuthState> {
         user: user,
       );
     } catch (error) {
+      final errorCode = error is AuthUiException ? error.code : null;
       state = state.copyWith(
         isLoading: false,
         initialized: true,
         isSignedIn: false,
         errorMessage: error.toString(),
+        errorCode: errorCode,
+        resetErrorCode: errorCode == null,
+        resetAwaitingConfirmation: true,
       );
       rethrow;
     }
@@ -119,32 +174,77 @@ class AuthController extends StateNotifier<AuthState> {
 
   Future<void> signUpWithEmail({
     required String username,
-    required String displayName,
     required String email,
     required String password,
   }) async {
-    state = state.copyWith(isLoading: true, errorMessage: null);
+    state = state.copyWith(
+      isLoading: true,
+      resetErrorMessage: true,
+      resetErrorCode: true,
+      resetAwaitingConfirmation: true,
+    );
     try {
-      await _authService.signUpEmail(
-        username: username,
-        displayName: displayName,
+      final normalized = _authService.normalizeUsername(username);
+      final result = await _authService.signUpEmail(
+        username: normalized,
         email: email,
         password: password,
       );
-      await signInWithEmail(usernameOrEmail: username, password: password);
+      if (result.isComplete || result.nextStep == 'none') {
+        await signInWithEmail(
+          usernameOrEmail: normalized,
+          password: password,
+        );
+        return;
+      }
+      if (result.nextStep == 'confirmCode') {
+        state = AuthState(
+          isLoading: false,
+          initialized: true,
+          isSignedIn: false,
+          user: null,
+          errorMessage: null,
+          awaitingConfirmation: AwaitingConfirmationState(
+            username: result.username,
+            email: email,
+            password: password,
+            deliveryDestination: result.deliveryDestination,
+            codeSentAt: DateTime.now(),
+          ),
+        );
+        return;
+      }
+      // Fallback: treat other next steps as incomplete but without follow-up.
+      state = state.copyWith(
+        isLoading: false,
+        initialized: true,
+        isSignedIn: false,
+        errorMessage:
+            'Sign-up requires additional steps. Please check your email for instructions.',
+        resetErrorCode: true,
+      );
     } catch (error) {
+      final errorCode = error is AuthUiException ? error.code : null;
       state = state.copyWith(
         isLoading: false,
         initialized: true,
         isSignedIn: false,
         errorMessage: error.toString(),
+        errorCode: errorCode,
+        resetErrorCode: errorCode == null,
+        resetAwaitingConfirmation: true,
       );
       rethrow;
     }
   }
 
   Future<void> signInWithGoogle() async {
-    state = state.copyWith(isLoading: true, errorMessage: null);
+    state = state.copyWith(
+      isLoading: true,
+      resetErrorMessage: true,
+      resetErrorCode: true,
+      resetAwaitingConfirmation: true,
+    );
     try {
       await _authService.signInWithGoogle();
       final user = await _authService.currentUser();
@@ -155,22 +255,126 @@ class AuthController extends StateNotifier<AuthState> {
         user: user,
       );
     } catch (error) {
+      final errorCode = error is AuthUiException ? error.code : null;
       state = state.copyWith(
         isLoading: false,
         initialized: true,
         isSignedIn: false,
         errorMessage: error.toString(),
+        errorCode: errorCode,
+        resetErrorCode: errorCode == null,
+        resetAwaitingConfirmation: true,
       );
       rethrow;
     }
   }
 
   Future<void> signOut() async {
-    await _authService.signOut();
+    state = state.copyWith(
+      isLoading: true,
+      resetErrorMessage: true,
+      resetErrorCode: true,
+      resetAwaitingConfirmation: true,
+    );
+    try {
+      await _authService.signOut();
+    } catch (error) {
+      state = state.copyWith(
+        isLoading: false,
+        errorMessage: error.toString(),
+        resetErrorCode: true,
+        resetAwaitingConfirmation: true,
+      );
+      return;
+    }
     state = AuthState(
       isLoading: false,
       initialized: true,
       isSignedIn: false,
+    );
+  }
+
+  Future<void> confirmSignUp({
+    required String username,
+    required String code,
+  }) async {
+    final awaiting = state.awaitingConfirmation;
+    state = state.copyWith(
+      isLoading: true,
+      resetErrorMessage: true,
+      resetErrorCode: true,
+    );
+    try {
+      await _authService.confirmSignUp(username: username, code: code);
+      final password = awaiting?.password;
+      if (password != null) {
+        await _authService.signInEmail(
+          usernameOrEmail: username,
+          password: password,
+        );
+        final user = await _authService.currentUser();
+        state = AuthState(
+          isLoading: false,
+          initialized: true,
+          isSignedIn: true,
+          user: user,
+        );
+      } else {
+        state = state.copyWith(
+          isLoading: false,
+          initialized: true,
+          isSignedIn: false,
+          resetErrorMessage: true,
+          resetErrorCode: true,
+          resetAwaitingConfirmation: true,
+        );
+      }
+    } catch (error) {
+      final errorCode = error is AuthUiException ? error.code : null;
+      state = state.copyWith(
+        isLoading: false,
+        initialized: true,
+        isSignedIn: false,
+        errorMessage: error.toString(),
+        errorCode: errorCode,
+        resetErrorCode: errorCode == null,
+      );
+      rethrow;
+    }
+  }
+
+  Future<void> resendConfirmationCode() async {
+    final awaiting = state.awaitingConfirmation;
+    if (awaiting == null) {
+      return;
+    }
+    state = state.copyWith(
+      isLoading: true,
+      resetErrorMessage: true,
+      resetErrorCode: true,
+    );
+    try {
+      await _authService.resendSignUpCode(username: awaiting.username);
+      state = state.copyWith(
+        isLoading: false,
+        awaitingConfirmation: awaiting.copyWith(codeSentAt: DateTime.now()),
+        resetErrorMessage: true,
+        resetErrorCode: true,
+      );
+    } catch (error) {
+      state = state.copyWith(
+        isLoading: false,
+        errorMessage: error.toString(),
+        resetErrorCode: true,
+      );
+      rethrow;
+    }
+  }
+
+  void clearError() {
+    state = state.copyWith(
+      resetErrorMessage: true,
+      resetErrorCode: true,
     );
   }
 }
