@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
+import '../../../providers/app_providers.dart';
+import '../../../services/auth_service.dart';
 import '../providers/auth_state.dart';
 
 class AuthGatePage extends ConsumerStatefulWidget {
@@ -26,14 +28,13 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
 
   final TextEditingController _signUpUsernameController =
       TextEditingController();
-  final TextEditingController _signUpDisplayNameController =
-      TextEditingController();
   final TextEditingController _signUpEmailController = TextEditingController();
   final TextEditingController _signUpPasswordController =
       TextEditingController();
 
   bool _obscureSignInPassword = true;
   bool _obscureSignUpPassword = true;
+  bool _isPresentingConfirmation = false;
 
   @override
   void initState() {
@@ -49,6 +50,14 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
             }
           });
         }
+        if (next.awaitingConfirmation != null &&
+            previous?.awaitingConfirmation == null) {
+          _presentConfirmation(next.awaitingConfirmation!);
+        }
+        if (next.awaitingConfirmation == null &&
+            previous?.awaitingConfirmation != null) {
+          _isPresentingConfirmation = false;
+        }
       },
       fireImmediately: false,
     );
@@ -57,6 +66,13 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
       WidgetsBinding.instance.addPostFrameCallback((_) {
         if (mounted) {
           context.go('/feed');
+        }
+      });
+    }
+    if (initialState.awaitingConfirmation != null) {
+      WidgetsBinding.instance.addPostFrameCallback((_) {
+        if (mounted && initialState.awaitingConfirmation != null) {
+          _presentConfirmation(initialState.awaitingConfirmation!);
         }
       });
     }
@@ -69,16 +85,98 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
     _signInUsernameController.dispose();
     _signInPasswordController.dispose();
     _signUpUsernameController.dispose();
-    _signUpDisplayNameController.dispose();
     _signUpEmailController.dispose();
     _signUpPasswordController.dispose();
     super.dispose();
+  }
+
+  void _presentConfirmation(AwaitingConfirmationState awaiting) {
+    if (_isPresentingConfirmation) {
+      return;
+    }
+    _isPresentingConfirmation = true;
+    final messenger = ScaffoldMessenger.of(context);
+    final destination = awaiting.deliveryDestination ?? awaiting.email;
+    messenger
+      ..hideCurrentSnackBar()
+      ..showSnackBar(
+        SnackBar(
+          content: Text(
+            'We sent a code to $destination. Enter it to finish signing up.',
+          ),
+        ),
+      );
+    WidgetsBinding.instance.addPostFrameCallback((_) async {
+      if (!mounted) {
+        _isPresentingConfirmation = false;
+        return;
+      }
+      try {
+        await context.push('/auth/confirm-code');
+      } finally {
+        _isPresentingConfirmation = false;
+      }
+    });
+  }
+
+  Widget? _buildStatusBanner(AuthState authState) {
+    final awaiting = authState.awaitingConfirmation;
+    if (authState.errorCode == 'username-exists') {
+      final message = authState.errorMessage ??
+          'An account with that username already exists. Try signing in or pick another handle.';
+      return Card(
+        margin: const EdgeInsets.symmetric(horizontal: 24),
+        color: Colors.orange.shade50,
+        child: Padding(
+          padding: const EdgeInsets.all(16),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                message,
+                style: TextStyle(
+                  color: Colors.orange.shade900,
+                  fontWeight: FontWeight.w600,
+                ),
+              ),
+              const SizedBox(height: 8),
+              Align(
+                alignment: Alignment.centerRight,
+                child: TextButton(
+                  onPressed: () {
+                    ref.read(authStateProvider.notifier).clearError();
+                    final username = _signUpUsernameController.text.trim();
+                    if (username.isNotEmpty) {
+                      _signInUsernameController.text = username;
+                    }
+                    _tabController.animateTo(0);
+                  },
+                  child: const Text('Go to Sign In'),
+                ),
+              ),
+            ],
+          ),
+        ),
+      );
+    }
+    if (authState.errorMessage != null && awaiting == null) {
+      return Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 24),
+        child: Text(
+          authState.errorMessage!,
+          style: const TextStyle(color: Colors.redAccent),
+          textAlign: TextAlign.center,
+        ),
+      );
+    }
+    return null;
   }
 
   @override
   Widget build(BuildContext context) {
     final authState = ref.watch(authStateProvider);
     final isLoading = authState.isLoading;
+    final statusBanner = _buildStatusBanner(authState);
 
     return Scaffold(
       body: SafeArea(
@@ -92,6 +190,10 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
                   style: TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
                 ),
                 const SizedBox(height: 16),
+                if (statusBanner != null) ...[
+                  statusBanner,
+                  const SizedBox(height: 16),
+                ],
                 TabBar(
                   controller: _tabController,
                   tabs: const [
@@ -136,6 +238,8 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
               decoration: const InputDecoration(
                 labelText: 'Username or email',
               ),
+              onChanged: (_) =>
+                  ref.read(authStateProvider.notifier).clearError(),
               validator: (value) {
                 if (value == null || value.trim().isEmpty) {
                   return 'Please enter your username or email';
@@ -163,6 +267,8 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
                 ),
               ),
               obscureText: _obscureSignInPassword,
+              onChanged: (_) =>
+                  ref.read(authStateProvider.notifier).clearError(),
               validator: (value) {
                 if (value == null || value.isEmpty) {
                   return 'Please enter your password';
@@ -192,27 +298,20 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
           crossAxisAlignment: CrossAxisAlignment.stretch,
           children: [
             TextFormField(
-              controller: _signUpDisplayNameController,
-              decoration: const InputDecoration(
-                labelText: 'Display name',
-              ),
-              validator: (value) {
-                if (value == null || value.trim().isEmpty) {
-                  return 'Display name is required';
-                }
-                return null;
-              },
-              textInputAction: TextInputAction.next,
-            ),
-            const SizedBox(height: 16),
-            TextFormField(
               controller: _signUpUsernameController,
               decoration: const InputDecoration(
                 labelText: 'Username',
               ),
+              onChanged: (_) =>
+                  ref.read(authStateProvider.notifier).clearError(),
               validator: (value) {
                 if (value == null || value.trim().isEmpty) {
                   return 'Username is required';
+                }
+                try {
+                  ref.read(authServiceProvider).normalizeUsername(value);
+                } catch (error) {
+                  return error.toString();
                 }
                 return null;
               },
@@ -221,6 +320,8 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
             TextFormField(
               controller: _signUpEmailController,
               decoration: const InputDecoration(labelText: 'Email'),
+              onChanged: (_) =>
+                  ref.read(authStateProvider.notifier).clearError(),
               validator: (value) {
                 if (value == null || value.trim().isEmpty) {
                   return 'Email is required';
@@ -251,6 +352,8 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
                 ),
               ),
               obscureText: _obscureSignUpPassword,
+              onChanged: (_) =>
+                  ref.read(authStateProvider.notifier).clearError(),
               validator: (value) {
                 if (value == null || value.length < 8) {
                   return 'Password must be at least 8 characters';
@@ -308,15 +411,20 @@ class _AuthGatePageState extends ConsumerState<AuthGatePage>
     try {
       await ref.read(authStateProvider.notifier).signUpWithEmail(
             username: _signUpUsernameController.text.trim(),
-            displayName: _signUpDisplayNameController.text.trim(),
             email: _signUpEmailController.text.trim(),
             password: _signUpPasswordController.text,
           );
     } catch (error) {
+      if (error is AuthUiException && error.code == 'username-exists') {
+        return;
+      }
+      final message = error is AuthUiException
+          ? (error.message ?? 'Sign-up failed. Please try again.')
+          : 'Sign-up failed: $error';
       messenger
         ..hideCurrentSnackBar()
         ..showSnackBar(
-          SnackBar(content: Text('Sign-up failed: $error')),
+          SnackBar(content: Text(message)),
         );
     }
   }
