@@ -1,59 +1,119 @@
+import 'dart:async';
+
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:flutter/foundation.dart';
-import 'package:google_sign_in/google_sign_in.dart';
 
 import '../amplifyconfiguration.dart';
 import '../env.dart';
 import '../features/auth/models/user_summary.dart';
 import 'session_manager.dart';
 
+class AuthFlowException implements Exception {
+  AuthFlowException(this.message, {this.cause});
+
+  final String message;
+  final Object? cause;
+
+  @override
+  String toString() => message;
+}
+
+class AuthUiException implements Exception {
+  AuthUiException(this.code, {this.message, this.cause});
+
+  final String code;
+  final String? message;
+  final Object? cause;
+
+  @override
+  String toString() => message ?? code;
+}
+
+class SignUpFlowResult {
+  SignUpFlowResult({
+    required this.isComplete,
+    required this.nextStep,
+    required this.username,
+    this.deliveryDestination,
+  });
+
+  final bool isComplete;
+  final String nextStep; // 'none' | 'confirmCode'
+  final String? deliveryDestination;
+  final String username;
+}
+
 class AuthService {
-  AuthService({SessionManager? sessionManager, GoogleSignIn? googleSignIn})
-      : _sessionManager = sessionManager ?? SessionManager(),
-        _googleSignIn = googleSignIn ?? GoogleSignIn(scopes: const ['email', 'profile']);
+  AuthService({SessionManager? sessionManager})
+      : _sessionManager = sessionManager ?? SessionManager();
 
   final SessionManager _sessionManager;
-  final GoogleSignIn _googleSignIn;
 
   bool _pluginsAdded = false;
-  bool _configured = false;
-  bool _isConfiguring = false;
+  Future<void>? _configureOperation;
 
-  Future<void> _configureIfNeeded() async {
+  static final RegExp _usernameRegExp = RegExp(r'^[a-z0-9_]{3,20}$');
+
+  Future<void> configureIfNeeded() async {
     if (kAuthBypassEnabled) {
       return;
     }
-    if (_configured) {
-      return;
-    }
-    if (_isConfiguring) {
-      while (_isConfiguring) {
-        await Future<void>.delayed(const Duration(milliseconds: 50));
-      }
-      return;
-    }
-    _isConfiguring = true;
-    try {
-      if (!_pluginsAdded) {
-        try {
-          Amplify.addPlugin(AmplifyAuthCognito());
-        } on Exception catch (error, stackTrace) {
-          debugPrint('[AuthService] addPlugin failed: $error\n$stackTrace');
-          rethrow;
-        }
-        _pluginsAdded = true;
-      }
 
-      try {
-        await Amplify.configure(amplifyconfig);
-      } on AmplifyAlreadyConfiguredException {
-        debugPrint('[AuthService] Amplify already configured; continuing');
-      }
-      _configured = true;
-    } finally {
-      _isConfiguring = false;
+    if (Amplify.isConfigured) {
+      return;
     }
+
+    if (_configureOperation != null) {
+      await _configureOperation;
+      return;
+    }
+
+    final completer = Completer<void>();
+    _configureOperation = completer.future;
+
+    try {
+      await _addPluginIfNeeded();
+      if (!Amplify.isConfigured) {
+        await Amplify.configure(amplifyconfig);
+      }
+      completer.complete();
+    } on AmplifyAlreadyConfiguredException {
+      debugPrint('[AuthService] Amplify already configured; continuing');
+      completer.complete();
+    } catch (error, stackTrace) {
+      debugPrint('[AuthService] configure failed: $error\n$stackTrace');
+      completer.completeError(error, stackTrace);
+      rethrow;
+    } finally {
+      _configureOperation = null;
+    }
+  }
+
+  Future<void> _addPluginIfNeeded() async {
+    if (_pluginsAdded) {
+      return;
+    }
+    try {
+      Amplify.addPlugin(AmplifyAuthCognito());
+      _pluginsAdded = true;
+    } on AmplifyAlreadyConfiguredException {
+      debugPrint('[AuthService] Auth plugin already added; continuing');
+      _pluginsAdded = true;
+    } on Exception catch (error, stackTrace) {
+      debugPrint('[AuthService] addPlugin failed: $error\n$stackTrace');
+      rethrow;
+    }
+  }
+
+  String normalizeUsername(String username) {
+    final normalized = username.trim().toLowerCase();
+    if (!_usernameRegExp.hasMatch(normalized)) {
+      throw AuthFlowException(
+        'Username must be 3-20 characters using lowercase letters, numbers, or underscores.',
+      );
+    }
+    return normalized;
   }
 
   Future<bool> isSignedIn() async {
@@ -61,13 +121,15 @@ class AuthService {
       final marker = await _sessionManager.readSessionMarker();
       return marker != null && marker.isNotEmpty;
     }
-    await _configureIfNeeded();
+    await configureIfNeeded();
     try {
       final session = await Amplify.Auth.fetchAuthSession();
       return session.isSignedIn;
+    } on SignedOutException {
+      return false;
     } on AuthException catch (error, stackTrace) {
       debugPrint('[AuthService] fetchAuthSession failed: $error\n$stackTrace');
-      return false;
+      throw _mapAuthException(error);
     }
   }
 
@@ -80,7 +142,7 @@ class AuthService {
       return UserSummary(userId: marker, username: marker, displayName: 'Developer');
     }
 
-    await _configureIfNeeded();
+    await configureIfNeeded();
     try {
       final user = await Amplify.Auth.getCurrentUser();
       final attributes = await Amplify.Auth.fetchUserAttributes();
@@ -106,7 +168,7 @@ class AuthService {
       );
     } on AuthException catch (error, stackTrace) {
       debugPrint('[AuthService] currentUser failed: $error\n$stackTrace');
-      return null;
+      throw _mapAuthException(error);
     }
   }
 
@@ -116,53 +178,109 @@ class AuthService {
       return;
     }
 
-    await _configureIfNeeded();
+    await configureIfNeeded();
     try {
-      await _googleSignIn.signOut();
-      final account = await _googleSignIn.signIn();
-      if (account == null) {
-        throw Exception('Google sign-in canceled');
-      }
-      await account.authentication;
       await Amplify.Auth.signInWithWebUI(provider: AuthProvider.google);
     } on AuthException catch (error, stackTrace) {
       debugPrint('[AuthService] Google sign-in failed: $error\n$stackTrace');
-      rethrow;
+      throw _mapAuthException(error);
     } catch (error, stackTrace) {
       debugPrint('[AuthService] Google sign-in unexpected error: $error\n$stackTrace');
-      rethrow;
+      throw AuthFlowException(
+        'Google sign-in failed. Please try again in a moment.',
+        cause: error,
+      );
     }
     await _persistSessionMarkerFromUser();
   }
 
-  Future<void> signUpEmail({
+  Future<SignUpFlowResult> signUpEmail({
     required String username,
-    required String displayName,
     required String email,
     required String password,
   }) async {
+    final normalizedUsername = normalizeUsername(username);
+
     if (kAuthBypassEnabled) {
-      await _sessionManager.persistSessionMarker(username);
-      return;
+      await _sessionManager.persistSessionMarker(normalizedUsername);
+      return SignUpFlowResult(
+        isComplete: true,
+        nextStep: 'none',
+        username: normalizedUsername,
+      );
     }
 
-    await _configureIfNeeded();
+    await configureIfNeeded();
     try {
-      final options = CognitoSignUpOptions(
+      final options = SignUpOptions(
         userAttributes: {
-          CognitoUserAttributeKey.email: email,
-          CognitoUserAttributeKey.preferredUsername: username,
-          CognitoUserAttributeKey.name: displayName,
+          AuthUserAttributeKey.email: email,
+          AuthUserAttributeKey.preferredUsername: normalizedUsername,
         },
       );
-      await Amplify.Auth.signUp(
-        username: username,
+      final result = await Amplify.Auth.signUp(
+        username: normalizedUsername,
         password: password,
         options: options,
       );
+      if (result.isSignUpComplete) {
+        return SignUpFlowResult(
+          isComplete: true,
+          nextStep: 'none',
+          username: normalizedUsername,
+        );
+      }
+      if (result.nextStep.signUpStep == AuthSignUpStep.confirmSignUp) {
+        final destination = result.nextStep.codeDeliveryDetails?.destination;
+        return SignUpFlowResult(
+          isComplete: false,
+          nextStep: 'confirmCode',
+          username: normalizedUsername,
+          deliveryDestination: destination,
+        );
+      }
+      final fallbackDestination = result.nextStep.codeDeliveryDetails?.destination;
+      return SignUpFlowResult(
+        isComplete: result.isSignUpComplete,
+        nextStep: 'none',
+        username: normalizedUsername,
+        deliveryDestination: fallbackDestination,
+      );
+    } on UsernameExistsException catch (error, stackTrace) {
+      debugPrint('[AuthService] signUpEmail failed: $error\n$stackTrace');
+      throw AuthUiException(
+        'username-exists',
+        message:
+            'An account with that username already exists. Try signing in or pick another handle.',
+        cause: error,
+      );
+    } on InvalidPasswordException catch (error, stackTrace) {
+      debugPrint('[AuthService] signUpEmail failed: $error\n$stackTrace');
+      throw AuthUiException(
+        'invalid-password',
+        message:
+            'Password does not meet policy requirements. Create a stronger password and try again.',
+        cause: error,
+      );
+    } on InvalidParameterException catch (error, stackTrace) {
+      debugPrint('[AuthService] signUpEmail failed: $error\n$stackTrace');
+      throw AuthUiException(
+        'invalid-parameter',
+        message:
+            'We could not process that sign-up request. Double-check your details and try again.',
+        cause: error,
+      );
+    } on CodeDeliveryFailureException catch (error, stackTrace) {
+      debugPrint('[AuthService] signUpEmail failed: $error\n$stackTrace');
+      throw AuthUiException(
+        'code-delivery-failed',
+        message:
+            'We could not send a verification code right now. Wait a moment and try again.',
+        cause: error,
+      );
     } on AuthException catch (error, stackTrace) {
       debugPrint('[AuthService] signUpEmail failed: $error\n$stackTrace');
-      rethrow;
+      throw _mapAuthException(error);
     }
   }
 
@@ -175,20 +293,60 @@ class AuthService {
       return;
     }
 
-    await _configureIfNeeded();
+    await configureIfNeeded();
+    final identifier = usernameOrEmail.contains('@')
+        ? usernameOrEmail.trim()
+        : normalizeUsername(usernameOrEmail);
     try {
       final result = await Amplify.Auth.signIn(
-        username: usernameOrEmail,
+        username: identifier,
         password: password,
       );
       if (!result.isSignedIn) {
-        throw Exception('Sign-in not completed');
+        throw AuthFlowException('Sign-in not completed. Please follow the next steps.');
       }
     } on AuthException catch (error, stackTrace) {
       debugPrint('[AuthService] signInEmail failed: $error\n$stackTrace');
-      rethrow;
+      throw _mapAuthException(error);
     }
     await _persistSessionMarkerFromUser();
+  }
+
+  Future<void> confirmSignUp({
+    required String username,
+    required String code,
+  }) async {
+    if (kAuthBypassEnabled) {
+      await _sessionManager.persistSessionMarker(username);
+      return;
+    }
+
+    await configureIfNeeded();
+    final normalizedUsername = normalizeUsername(username);
+    try {
+      await Amplify.Auth.confirmSignUp(
+        username: normalizedUsername,
+        confirmationCode: code.trim(),
+      );
+    } on AuthException catch (error, stackTrace) {
+      debugPrint('[AuthService] confirmSignUp failed: $error\n$stackTrace');
+      throw _mapAuthException(error);
+    }
+  }
+
+  Future<void> resendSignUpCode({required String username}) async {
+    if (kAuthBypassEnabled) {
+      return;
+    }
+
+    await configureIfNeeded();
+    final normalizedUsername = normalizeUsername(username);
+    try {
+      await Amplify.Auth.resendSignUpCode(username: normalizedUsername);
+    } on AuthException catch (error, stackTrace) {
+      debugPrint('[AuthService] resendSignUpCode failed: $error\n$stackTrace');
+      throw _mapAuthException(error);
+    }
   }
 
   Future<void> signOut() async {
@@ -197,11 +355,14 @@ class AuthService {
       return;
     }
 
-    await _configureIfNeeded();
+    await configureIfNeeded();
     try {
-      await Amplify.Auth.signOut();
+      await Amplify.Auth.signOut(
+        options: const SignOutOptions(globalSignOut: true),
+      );
     } on AuthException catch (error, stackTrace) {
       debugPrint('[AuthService] signOut failed: $error\n$stackTrace');
+      throw _mapAuthException(error);
     }
     await _sessionManager.clearSessionMarker();
   }
@@ -210,15 +371,15 @@ class AuthService {
     if (kAuthBypassEnabled) {
       return null;
     }
-    await _configureIfNeeded();
+    await configureIfNeeded();
     try {
       final session = await Amplify.Auth.fetchAuthSession(
         options: const FetchAuthSessionOptions(forceRefresh: false),
       );
       if (session is CognitoAuthSession) {
-        final result = session.userPoolTokensResult;
-        if (result.value != null) {
-          return result.value.idToken;
+        final tokens = session.userPoolTokensResult.valueOrNull;
+        if (tokens != null) {
+          return tokens.idToken.raw;
         }
       }
     } on AuthException catch (error, stackTrace) {
@@ -233,5 +394,46 @@ class AuthService {
     if (marker != null && marker.isNotEmpty) {
       await _sessionManager.persistSessionMarker(marker);
     }
+  }
+
+  AuthFlowException _mapAuthException(AuthException error) {
+    if (error is UsernameExistsException) {
+      return AuthFlowException(
+        'An account with that username already exists. Try signing in or pick another handle.',
+        cause: error,
+      );
+    }
+    if (error is AuthNotAuthorizedException || error is InvalidPasswordException) {
+      return AuthFlowException(
+        'Incorrect username or password. Please try again.',
+        cause: error,
+      );
+    }
+    if (error is UserNotFoundException) {
+      return AuthFlowException(
+        'We couldn\'t find an account with those details. Check the spelling or sign up.',
+        cause: error,
+      );
+    }
+    if (error is CodeMismatchException || error is ExpiredCodeException) {
+      return AuthFlowException(
+        'The verification code is incorrect or expired. Request a new code and try again.',
+        cause: error,
+      );
+    }
+    if (error.message.toLowerCase().contains('not configured')) {
+      return AuthFlowException(
+        'Auth not configured. Rebuild after running amplify pull.',
+        cause: error,
+      );
+    }
+    if (error is NetworkException ||
+        error.recoverySuggestion?.toLowerCase().contains('network') == true) {
+      return AuthFlowException(
+        'We hit a network issue. Check your connection and try again.',
+        cause: error,
+      );
+    }
+    return AuthFlowException(error.message, cause: error);
   }
 }
